Update to V1.0.0 (#270)
* update readme

* fix gt_size bug during test in paired_image_dataset

* update workflows

* update .gitignore

* add datasets README

* move convert_models and update_model

* update setup

* update readme

* update readme

* update .gitignore

* remove seaborn package

* update readme

* rm meta info

* update generate_meta_info

* update num_gpu

* update about rank, seed, config, etc

* update EnlargedSampler

* update ffhq_dataset

* update augment

* update PairedImageDataset

* update PairedImageDataset

* replace ann_file with meta_info_file

* update reds dataset

* make 'scale' optional in options

* update data util

* update video_test_dataset

* update vimeo90k_dataset

* update metric_util

* update metric

* rm test_Vid4_REDS4_with_GT

* update logger

* use total_iter

* use result

* update options

* update util

* update setup.py

* update train.py

* update train and test

* add download

* add download pretrained models

* update readme

* remove unnecssary int

* use training_state

* update loss_utils

* update pixel-wise losses

* update perceptual losses

* update gan loss

* update ops

* arch utils

* update archs

* update base model

* update models

* update read me

* update train opt

* update test options

* update readme

* worker_init_fn only work in train phase

* update dataset_enlarge_ratio option

* pin_memory must be True for CUDAPrefetcher

* only make dirs for logger in rank 0

* fix tsa_iter bug

* only make dirs for logger in rank 0

* fix get_dist_info bug

* update dataset_enlarge_ratio

* data lack of test

* add img_gt.squeeze_

* set shuffle=True when sampler=None

* update test_ffhq

* update publish_models

* update publish_models

* update readme

* update LICENSE

* update readme

Former-commit-id: 06fb33acf5d138615c8c3153ee5872550d3197ca

+++ basicsr/models/archs/spynet_arch.py
@@ -130,8 +130,8 @@ class SpyNet(nn.Module):
         assert ref.size() == supp.size()
 
         h, w = ref.size(2), ref.size(3)
-        w_floor = int(math.floor(math.ceil(w / 32.0) * 32.0))
-        h_floor = int(math.floor(math.ceil(h / 32.0) * 32.0))
+        w_floor = math.floor(math.ceil(w / 32.0) * 32.0)
+        h_floor = math.floor(math.ceil(h / 32.0) * 32.0)
 
         ref = F.interpolate(
             input=ref,
